Use a single long value on BitsValue

+++ LogicScript/Parsing/Structures/NumberLiteralExpression.cs
@@ -2,12 +2,12 @@
 {
     internal class NumberLiteralExpression : Expression
     {
-        public int Value { get; }
+        public long Value { get; }
         public int? Length { get; }
 
         public override bool IsSingleBit => Length == 1;
 
-        public NumberLiteralExpression(SourceLocation location, int value, int? length = null) : base(location)
+        public NumberLiteralExpression(SourceLocation location, long value, int? length = null) : base(location)
         {
             this.Value = value;
             this.Length = length;
